requirements.txt
openai
python-dotenv
langchain
chromadb
flask_httpauth
flask_cors
jwt

